+++ src/main/java/map/homepage/domain/post/comment/Dto/CommentCreateRequest.java
@@ -1,4 +1,4 @@
-package map.homepage.domain.post.comment;
+package map.homepage.domain.post.comment.Dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
